Remove SpecialistDocumentEdition#extra_fields field
This field was only ever used for non-manual specialist documents,
but these are now stored in the Specialist Publisher app. So the
field is no longer needed and can be removed.

There hasn't been a way to set this field via the user interface
since the giant "Remove specialist document models" commit [1].

I've run the following queries against the production data dumped on
2017-02-24 to confirm that there are no documents with this field
set:

    > SpecialistDocumentEdition.where(extra_fields: nil).count
    => 0
    > SpecialistDocumentEdition.where(extra_fields: []).count
    => 0

The field is still referenced in some old migrations, but don't think
there's any need to modify them.

[1]: 3162f0e72fd53bfd1cf87f5661bb198ff785940e